bird-detection/requirements.txt
# Function dependencies, for example:
# package>=version
opencv-python==4.2.0.34
numpy==1.18.2
google-cloud-storage==1.27.0
tensorflow==2.3.0
Pillow>=0.0.1
keras.models>=0.0.1
drive>==0.0.1
h5py==2.10.0
gcsfs>==0.0.1
argparse>==0.0.1

